pyproject.toml
[project]
name = "sai-intw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pdfminer.six",
    "requests",
    "numpy",
    "python-multipart"
]
